pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "buddy-agent"
version = "0.1.0"
description = "Buddy Agent - Process Document Q&A Assistant"
authors = [
    {name = "Google LLC", email = "adk-samples@google.com"}
]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "google-adk>=0.1.0",
    "google-cloud-aiplatform>=1.38.0",
    "google-genai>=0.3.0",
    "pdfplumber>=0.9.0",
    "python-docx>=0.8.11",
    "pandas>=1.5.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/google/adk-samples"
Repository = "https://github.com/google/adk-samples"
Documentation = "https://cloud.google.com/vertex-ai/generative-ai/docs/agent-engine/develop/adk"

[tool.setuptools.packages.find]
where = ["."]
include = ["buddy_agent*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

